refactor: 엔티티 명 변경에 따라 result에서 row로 변경
