+++ src/main/java/racingcar/RacingCarConsoleApplication.java
@@ -1,7 +1,7 @@
 package racingcar;
 
 import racingcar.controller.console.RacingGameConsoleController;
-import racingcar.dao.GameResultInMemoryDao;
+import racingcar.dao.gameresult.GameRowInMemoryDao;
 import racingcar.service.RacingCarService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -10,7 +10,7 @@ public final class RacingCarConsoleApplication {
 
 
 	public static void main (String[] args) {
-		final RacingGameConsoleController racingGameConsoleController = new RacingGameConsoleController(new InputView(), new OutputView(), new RacingCarService(new GameResultInMemoryDao()));
+		final RacingGameConsoleController racingGameConsoleController = new RacingGameConsoleController(new InputView(), new OutputView(), new RacingCarService(new GameRowInMemoryDao()));
 		racingGameConsoleController.run();
 	}
 }
