+++ src/main/java/racingcar/dao/gameresult/GameRowDao.java
@@ -1,12 +1,12 @@
-package racingcar.dao;
+package racingcar.dao.gameresult;
 
 import java.util.List;
 import racingcar.domain.TryCount;
 import racingcar.entity.GameRow;
 import racingcar.entity.PlayerRow;
 
-public interface GameResultDao {
+public interface GameRowDao {
 	void saveGame (final List<PlayerRow> playerRows, final TryCount tryCount);
-	List<GameRow> fetchAllGameResult ();
+	List<GameRow> fetchAllGameRow ();
 	List<PlayerRow> fetchAllPlayerResultByGameId (Long gameId);
 }
