+++ src/main/java/racingcar/service/RacingCarService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import racingcar.dao.GameResultDao;
+import racingcar.dao.gameresult.GameRowDao;
 import racingcar.domain.Cars;
 import racingcar.domain.TryCount;
 import racingcar.dto.CarStatusResponseDto;
@@ -19,11 +19,11 @@ import racingcar.utils.RandomPowerMaker;
 public class RacingCarService {
 
 	private final RandomPowerGenerator randomPowerGenerator;
-	private final GameResultDao gameResultDao;
+	private final GameRowDao gameRowDao;
 
-	public RacingCarService (final GameResultDao gameResultDao) {
+	public RacingCarService (final GameRowDao gameRowDao) {
 		this.randomPowerGenerator = new RandomPowerMaker();
-		this.gameResultDao = gameResultDao;
+		this.gameRowDao = gameRowDao;
 	}
 
 	@Transactional
@@ -33,7 +33,7 @@ public class RacingCarService {
 			cars.moveAll(randomPowerGenerator);
 		}
 		List<PlayerRow> playerRows = toPlayerResults(cars);
-		gameResultDao.saveGame(playerRows, tryCount);
+		gameRowDao.saveGame(playerRows, tryCount);
 
 		return GameResultResponseDto.toDto(cars.getWinnerNames(), cars);
 	}
@@ -48,7 +48,7 @@ public class RacingCarService {
 	}
 
 	public List<GameResultResponseDto> findAllGameResults () {
-		List<GameRow> gameRows = gameResultDao.fetchAllGameResult();
+		List<GameRow> gameRows = gameRowDao.fetchAllGameRow();
 		List<GameResultResponseDto> gameResultResponseDtos = new ArrayList<>();
 		for (final GameRow gameRow : gameRows) {
 			gameResultResponseDtos.add(gameResultInOneGame(gameRow));
@@ -58,7 +58,7 @@ public class RacingCarService {
 	}
 
 	private GameResultResponseDto gameResultInOneGame (final GameRow gameRow) {
-		List<PlayerRow> playerRows = gameResultDao.fetchAllPlayerResultByGameId(gameRow.getId());
+		List<PlayerRow> playerRows = gameRowDao.fetchAllPlayerResultByGameId(gameRow.getId());
 		List<String> winners = new ArrayList<>();
 		List<CarStatusResponseDto> carStatusResponseDtos = new ArrayList<>();
 		for (final PlayerRow playerRow : playerRows) {
